add screwdriver plugin
The screwdriver plugin is activated once a configurable button on the
wiimote (B by default) is pressed and held. Turning the wiimote
generates key events as long as the button is held down and the wiimote
is being turned. The threshold for indicating a new event is
configurable. The movement resembles the use of a screwdriver and can be
for example be used to regulate the volume of a media player system,
which feels natural and intuitive.